requirements.txt
absl-py==2.1.0
arabic-reshaper==3.0.0
asn1crypto==1.5.1
astunparse==1.6.3
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
contourpy==1.2.1
cryptography==42.0.5
cssselect2==0.7.0
cycler==0.12.1
et-xmlfile==1.1.0
flatbuffers==24.3.25
fonttools==4.50.0
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.10.0
html5lib==1.1
idna==3.6
keras==3.1.1
kiwisolver==1.4.5
libclang==18.1.1
lxml==5.2.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
mdurl==0.1.2
ml-dtypes==0.3.2
namex==0.0.7
numpy==1.26.4
opencv-python==4.9.0.80
openpyxl==3.1.2
opt-einsum==3.3.0
optree==0.11.0
oscrypto==1.3.0
packaging==24.0
pandas==2.2.0
pillow==10.2.0
protobuf==4.25.3
pyarrow==15.0.0
pycparser==2.22
Pygments==2.17.2
pyHanko==0.23.2
pyhanko-certvalidator==0.26.3
pyparsing==3.1.2
pypdf==4.1.0
pypng==0.20220715.0
python-bidi==0.4.2
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
pyzbar==0.1.9
qrcode==7.4.2
reportlab==4.0.9
requests==2.31.0
rich==13.7.1
six==1.16.0
svglib==1.5.1
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-intel==2.16.1
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
tinycss2==1.2.1
typing_extensions==4.10.0
tzdata==2024.1
tzlocal==5.2
uritools==4.0.2
urllib3==2.2.1
webencodings==0.5.1
Werkzeug==3.0.2
wrapt==1.16.0
xhtml2pdf==0.2.15
